+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.2.2
 
+* Fix issue with description type that did not exist.
+
 ## 1.2.1
 
 * [#25](https://github.com/brightin/poms/pull/25) Fix issue with `Timestamp.convert`